+++ stories/02-Design System/03-Navigation/Breadcrumb/examples/icon.tsx
@@ -3,20 +3,17 @@ import {KitBreadcrumb} from '@kit/Navigation';
 import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
 import {faHouse} from '@fortawesome/free-solid-svg-icons';
 import {faUser} from '@fortawesome/free-regular-svg-icons';
+import {IKitBreadcrumbItem} from '@kit/Navigation/Breadcrumb/types';
 
-const items = [
+const items: IKitBreadcrumbItem[] = [
     {
         href: '',
-        title: <FontAwesomeIcon icon={faHouse} />
+        icon: <FontAwesomeIcon icon={faHouse} />
     },
     {
         href: '',
-        title: (
-            <>
-                <FontAwesomeIcon icon={faUser} />
-                <span style={{marginLeft: '6px'}}>Application List</span>
-            </>
-        )
+        title: 'Application List',
+        icon: <FontAwesomeIcon icon={faUser} />
     },
     {
         title: 'Application'